+++ package.js
@@ -23,17 +23,16 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
-    api.versionsFrom( '2.9.0' );
-    api.use( 'alanning:roles@3.4.0' );
-    api.use( 'blaze-html-templates@2.0.0', 'client' );
-    api.use( 'deanius:promise@3.1.3', 'client' );
+    api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
+    api.use( 'alanning:roles@3.4.0 || 4.0.0-alpha.0' );
+    api.use( 'blaze-html-templates@2.0.0 || 3.0.0-alpha300.0', 'client' );
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'mongo', 'server' );
-    api.use( 'pwix:i18n@1.5.2' );
-    api.use( 'pwix:jstree@1.0.3' );
-    api.use( 'pwix:modal@1.7.1' );
-    api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
+    api.use( 'pwix:i18n@1.5.7' );
+    api.use( 'pwix:jstree@1.0.6' );
+    api.use( 'pwix:modal@1.10.0' );
+    api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.addFiles( 'src/client/components/prEdit/prEdit.js', 'client' );
     api.addFiles( 'src/client/components/prEditPanel/prEditPanel.js', 'client' );
     api.addFiles( 'src/client/components/prView/prView.js', 'client' );