+++ src/server/js/accounts.js
@@ -5,12 +5,12 @@
 Meteor.methods({
     // returns the requested user
     'Roles.Accounts.User'( id ){
-        return Meteor.users.findOne({ _id: id });
+        return Meteor.users.findOneAsync({ _id: id });
     },
 
     // update updatedAt / updatedBy when updating the user's roles
     'Roles.Accounts.Updated'( id ){
-        Meteor.users.update({ _id: id }, { $set: {
+        return Meteor.users.updateAsync({ _id: id }, { $set: {
             updatedAt: new Date(),
             updatedBy: Meteor.userId()
         }});