+++ src/server/js/functions.js
@@ -5,9 +5,10 @@
 import _ from 'lodash';
 
 Roles.server = {
-    // get roles for the users  
+    // get roles for the user
+    // return a Promise which resolves to the array of roles for the user (which may be empty)
     getRolesForUser( user, options ){
-        let result = [];
+        let result = Promise.resolve( [] );
         if( user ){
             let id = null;
             if( _.isString( user )){
@@ -16,16 +17,21 @@ Roles.server = {
                 id = user._id;
             }
             if( id ){
-                Meteor.roleAssignment.find({ 'user._id': id }).fetch().every(( doc ) => {
-                    if( options.onlyScoped === true ){
-                        if(( doc.scope && options.scope && doc.scope === options.scope ) || ( !options.scope && !doc.scope )){
-                            result.push({ _id: doc.role._id, scope: doc.scope || null });
-                        }
-                    } else {
-                        result.push({ _id: doc.role._id, scope: doc.scope || null });
-                    }
-                    return true;
-                });
+                return Meteor.roleAssignment.find({ 'user._id': id }).fetchAsync()
+                    .then(( fetched ) => {
+                        result = [];
+                        fetched.every(( doc ) => {
+                            if( options.onlyScoped === true ){
+                                if(( doc.scope && options.scope && doc.scope === options.scope ) || ( !options.scope && !doc.scope )){
+                                    result.push({ _id: doc.role._id, scope: doc.scope || null });
+                                }
+                            } else {
+                                result.push({ _id: doc.role._id, scope: doc.scope || null });
+                            }
+                            return true;
+                        });
+                        return result;
+                    });
             } else {
                 console.warn( 'getRolesForUser() unable to find an identifier', user );
             }
@@ -35,23 +41,29 @@ Roles.server = {
         return result;
     },
 
-    // get user in scope
+    // get users in scope
+    // returns a Promise which resolves to the list of users in this scope, or null
     getUsersInScope( scope ){
-        let ret = [];
+        let result = Promise.resolve( [] );
         if( scope && scope.length && _.isString( scope )){
-            const assigns = Meteor.roleAssignment.find({ scope: scope }).fetch();
-            assigns.every(( it ) => {
-                ret.push( user._id );
-                return true;
-            })
+            return Meteor.roleAssignment.find({ scope: scope }).fetchAsync()
+                .then(( fetched ) => {
+                    result = [];
+                    fetched.every(( it ) => {
+                        result.push( user._id );
+                        return true;
+                    });
+                    return result;
+                });
+            
         }
-        return ret;
+        return result;
     },
 
     // remove all roles for the user
-    //  returns true|false
+    //  returns a Promise which resolves to true|false
     removeAllRolesFromUser( user ){
-        let ret = -1;
+        let result = Promise.resolve( false );
         if( user ){
             let id = null;
             if( _.isString( user )){
@@ -60,21 +72,20 @@ Roles.server = {
                 id = user._id;
             }
             if( id ){
-                ret = Meteor.roleAssignment.remove({ 'user._id': id });
+                result = Meteor.roleAssignment.removeAsync({ 'user._id': id });
             } else {
                 console.warn( 'removeAllRolesFromUser() unable to find an identifier', user );
             }
         } else {
             console.warn( 'removeAllRolesFromUser() user is falsy', user );
         }
-        console.debug( 'removeAllRolesFromUser() ret=', ret )
-        return ret;
+        return result;
     },
 
     // remove all assignments for the role(s)
-    //  returns true|false
+    //  returns a Promise which resolves an array of the result for each role
     removeUserAssignmentsForRoles( roles, opts ){
-        let ret = -1;
+        let promises = [];
         const rolesArray = _.isArray( roles ) ? roles : [roles];
         rolesArray.every(( role ) => {
             let query = {
@@ -83,10 +94,10 @@ Roles.server = {
             if( opts.scope ){
                 query.scope = opts.scope;
             }
-            ret = Meteor.roleAssignment.remove( query );
+            promises.push( Meteor.roleAssignment.removeAsync( query ));
             //console.debug( 'removeUserAssignmentsForRoles()', 'query', query, 'ret', ret );
             return true;
         });
-        return ret;
+        return Promise.all( promises );
     }
 };