+++ src/server/js/methods.js
@@ -12,20 +12,20 @@ import { Roles as alRoles } from 'meteor/alanning:roles';
 Meteor.methods({
     // returns the count of users which have at least one of the specified roles
     'Roles.addUsersToRoles'( users, roles, options={} ){
-        alRoles.addUsersToRoles( users, roles, options );
+        alRoles.addUsersToRolesAsync( users, roles, options );
         console.log( 'pwix:roles/src/server/js/methods:addUsersToRoles()' );
     },
 
     // returns the count of users which have at least one of the specified roles
     'Roles.countUsersInRoles'( roles, options={} ){
-        const res = alRoles.getUsersInRole( roles, options ).count();
+        const res = alRoles.getUsersInRoleAsync( roles, options ).count();
         console.log( 'pwix:roles/src/server/js/methods:countUsersInRoles()', roles, res );
         return res;
     },
 
     // create a new role (when we do not want manage it in the hierarchy)
     'Roles.createRole'( role, options={} ){
-        const res = alRoles.createRole( role, options );
+        const res = alRoles.createRoleAsync( role, options );
         console.log( 'pwix:roles/src/server/js/methods:createRole()', res );
         return res;
     },
@@ -53,10 +53,11 @@ Meteor.methods({
     // replace the user's roles with a new set
     //  this must be a method as the (not trusted) client cannot directly remove assignments without its id
     'Roles.setUsersRoles'( users, roles ){
+        let promises = [];
         const ids = Roles.idsFromUsers( users );
         ids.every(( id ) => {
-            Meteor.roleAssignment.remove({ 'user._id': id });
+            promises.push( Meteor.roleAssignment.removeAsync({ 'user._id': id }));
         });
-        alRoles.addUsersToRoles( users, roles );
+        return Promise.all( promises ).then(() => { return alRoles.addUsersToRolesAsync( users, roles ); });
     }
 });