+++ src/server/js/publish.js
@@ -198,23 +198,25 @@ Meteor.publish( 'Roles.listByRole', function( roles ){
     let collectionHash = {};
 
     // `observeChanges` only returns after the initial `added` callbacks have run.
-    // Until then, we don't want to send a lot of `changed` messages—hence
-    // tracking the `initializing` state.
+    // Until then, we don't want to send a lot of `changed` messages
+    // hence tracking the `initializing` state.
 
-    const handle = alRoles.getUsersInRole( rolesArray ).observeChanges({
+    const handle = alRoles.getUsersInRoleAsync( rolesArray ).observeChanges({
         added( user_id, user_doc ){
             //console.debug( 'added', arguments );
-            const userRoles = alRoles.getRolesForUser( user_id );
-            rolesArray.every(( role ) => {
-                if( userRoles.includes( role )){
-                    if( Object.keys( collectionHash ).includes( role )){
-                        collectionHash[role] += 1;
-                    } else {
-                        collectionHash[role] = 1;
-                    }
-                }
-                return true;
-            });
+            alRoles.getRolesForUserAsync( user_id )
+                .then(( userRoles ) => {
+                    rolesArray.every(( role ) => {
+                        if( userRoles.includes( role )){
+                            if( Object.keys( collectionHash ).includes( role )){
+                                collectionHash[role] += 1;
+                            } else {
+                                collectionHash[role] = 1;
+                            }
+                        }
+                        return true;
+                    });
+                });
         },
         // cannot handle changed() nor removed() as obviously the roles have changed or have been removed
         changed(){